+++ src/client/pages/Registration.tsx
@@ -15,9 +15,8 @@ import { Terms } from '@/client/components/Terms';
 import { SocialButtons } from '@/client/components/SocialButtons';
 import { button, form, textInput } from '@/client/styles/Shared';
 import { css } from '@emotion/react';
-import { Breakpoints } from '../models/Style';
-import { brandLine, space } from '@guardian/src-foundations';
-import { recaptchaReady } from '../static/recaptcha/recaptcha';
+import { space } from '@guardian/src-foundations';
+import { RecaptchaElement, useRecaptcha } from '../static/recaptcha/recaptcha';
 
 type Props = {
   returnUrl?: string;
@@ -25,93 +24,9 @@ type Props = {
   recaptchaSiteKey?: string;
 };
 
-// Used to centre the nav tabs in line with the page content.
-const centreTabRow = css`
-  max-width: ${Breakpoints.TABLET}px;
-  width: 100%;
-  margin: 0 auto;
-  border-left: 1px solid ${brandLine.primary};
-`;
-
 const termsSpacing = css`
   margin-bottom: ${space[4]}px;
 `;
-
-const useRecaptchaScript = (url: string) => {
-  const [loaded, setLoaded] = React.useState(false);
-
-  React.useEffect(() => {
-    const existingScript =
-      typeof document !== undefined &&
-      document.getElementById('g-captcha-script');
-
-    if (existingScript || recaptchaReady()) {
-      setLoaded(true);
-      return;
-    }
-
-    const script = document.createElement('script');
-
-    script.setAttribute('src', url);
-
-    script.setAttribute('id', 'g-captcha-script');
-    script.setAttribute('async', '');
-    script.setAttribute('defer', '');
-
-    document.body.appendChild(script);
-
-    const initialiseRecaptcha = () => {
-      window.grecaptcha.ready(() => {
-        if (recaptchaReady()) {
-          setLoaded(true);
-        }
-      });
-    };
-    script.addEventListener('load', initialiseRecaptcha);
-
-    return () => {
-      script.removeEventListener('load', initialiseRecaptcha);
-      // TODO: Decide whether to remove from body on effect cleanup.
-      document.body.removeChild(script);
-    };
-  }, []);
-
-  return {
-    loaded,
-  };
-};
-
-const useRecaptcha = (siteKey: string, renderElement: string) => {
-  const [token, setToken] = React.useState('');
-  const [error, setError] = React.useState('');
-  const [expired, setExpired] = React.useState('');
-
-  const [widgetId, setWidgetId] = React.useState(0);
-
-  const { loaded } = useRecaptchaScript(
-    'https://www.google.com/recaptcha/api.js?render=explicit',
-  );
-
-  React.useEffect(() => {
-    if (loaded) {
-      const widgetId = window.grecaptcha.render(renderElement, {
-        sitekey: siteKey,
-        size: 'invisible',
-        callback: setToken,
-        'error-callback': setError,
-        'expired-callback': setExpired,
-      });
-      setWidgetId(widgetId);
-    }
-  }, [loaded]);
-
-  return {
-    token,
-    error,
-    expired,
-    captchaElement: widgetId,
-  };
-};
 
 export const Registration = ({
   returnUrl = '',
@@ -119,30 +34,28 @@ export const Registration = ({
   recaptchaSiteKey = '',
 }: Props) => {
   const registerFormRef = React.createRef<HTMLFormElement>();
-
-  const { token, captchaElement } = useRecaptcha(
-    recaptchaSiteKey,
-    'recaptcha_registration',
-  );
-
+  const recaptchaElementRef = React.useRef<HTMLDivElement>(null);
   const returnUrlQuery = returnUrl
     ? `?returnUrl=${encodeURIComponent(returnUrl)}`
     : '';
+
+  const { token, widgetId } = useRecaptcha(
+    recaptchaSiteKey,
+    recaptchaElementRef.current,
+  );
 
   React.useEffect(() => {
     const registerFormElement = registerFormRef.current;
     if (token) {
       console.log('Token rcv: ', token);
-      // registerFormElement.submit();
+      registerFormElement?.submit();
     }
   }, [token]);
 
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
-    if (captchaElement !== null) {
-      window.grecaptcha.reset(captchaElement);
-      window.grecaptcha.execute(captchaElement);
-    }
+    window.grecaptcha.reset(widgetId);
+    window.grecaptcha.execute(widgetId);
   };
 
   return (
@@ -161,7 +74,6 @@ export const Registration = ({
             isActive: true,
           },
         ]}
-        tabRowStylesOverride={centreTabRow}
       />
 
       <Main>
@@ -175,7 +87,7 @@ export const Registration = ({
               onSubmit={handleSubmit}
             >
               <CsrfFormField />
-              <div className="g-recaptcha" id="recaptcha_registration"></div>
+              <RecaptchaElement ref={recaptchaElementRef} />
               <TextInput
                 css={textInput}
                 label="Email"
