--- src/client/static/recaptcha/recaptcha.ts
@@ -1,4 +1,0 @@
-export const recaptchaReady = () =>
-  typeof window !== 'undefined' &&
-  typeof window.grecaptcha !== 'undefined' &&
-  typeof window.grecaptcha.render === 'function';
